refactor(settings): gather the cache sections inside the cache management feature
VoicemailCacheSection and CdrsCacheSection lived in the voicemail and
cdrs features; they implement the cache management contract and belong
to that feature, so they move into its sections/ folder.
